requirements.txt
lark-parser
Wand
numpy
matplotlib
ipython
